requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.9
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
websockets==11.0.2
pydantic==1.10.7
livekit==0.21.3
livekit-agents==0.12.17
jinja2==3.1.2
python-dotenv==1.0.0